python/requirements.txt
paho-mqtt
inputs
#xbox360controller